pyproject.toml
[project]
name = "event-calendar-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis]>=5.5.3",
    "channels>=4.3.1",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.11",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.5",
    "redis>=5.2.1",
    "requests>=2.32.5",
]
